Code polishing

+++ src/main/java/org/jsqrl/server/JSqrlServer.java
@@ -47,17 +47,17 @@ import java.util.Set;
 @Slf4j
 public class JSqrlServer {
 
-    private SqrlUserService userService;
-    private SqrlAuthenticationService sqrlAuthenticationService;
-    private SqrlConfig config;
-    private SqrlNutService nutService;
+    private final SqrlUserService userService;
+    private final SqrlAuthenticationService sqrlAuthenticationService;
+    private final SqrlConfig config;
+    private final SqrlNutService nutService;
 
-    private EdDSAParameterSpec edDsaSpec;
+    private final EdDSAParameterSpec edDsaSpec;
 
-    public JSqrlServer(SqrlUserService userService,
-                       SqrlAuthenticationService sqrlAuthenticationService,
-                       SqrlConfig config,
-                       SqrlNutService nutService) {
+    public JSqrlServer(final SqrlUserService userService,
+                       final SqrlAuthenticationService sqrlAuthenticationService,
+                       final SqrlConfig config,
+                       final SqrlNutService nutService) {
         this.userService = userService;
         this.sqrlAuthenticationService = sqrlAuthenticationService;
         this.config = config;
@@ -75,7 +75,7 @@ public class JSqrlServer {
      * @param qr        If the user used the QR code or not
      * @return Returns the nut to be provided to the user
      */
-    public String createAuthenticationRequest(String ipAddress, Boolean qr) {
+    public String createAuthenticationRequest(final String ipAddress, final Boolean qr) {
         SqrlNut nut = nutService.createNut(ipAddress, qr);
         String nutString = nutService.getNutString(nut);
         sqrlAuthenticationService.createAuthenticationRequest(nutString, ipAddress);
@@ -92,19 +92,13 @@ public class JSqrlServer {
      * @param ipAddress The requesting IP Address
      * @return Returns true if the nut was marked as authenticated
      */
-    public Boolean checkAuthenticationStatus(String nut, String ipAddress) {
-
-        if (!nutService.nutBelongsToIp(nut, ipAddress)) {
-            return false;
-        } else if (sqrlAuthenticationService.getAuthenticatedSqrlIdentityKey(nut, ipAddress) != null) {
-            return true;
-        } else return false;
-
+    public Boolean checkAuthenticationStatus(final String nut, final String ipAddress) {
+        return nutService.nutBelongsToIp(nut, ipAddress) && sqrlAuthenticationService.getAuthenticatedSqrlIdentityKey(nut, ipAddress) != null;
     }
 
-    public SqrlAuthResponse handleClientRequest(SqrlClientRequest request,
-                                                String nut,
-                                                String ipAddress) {
+    public SqrlAuthResponse handleClientRequest(final SqrlClientRequest request,
+                                                final String nut,
+                                                final String ipAddress) {
 
         //Build the new nut for this request, retain the QR code
         SqrlNut requestNut = nutService.createNutFromString(nut);
@@ -201,13 +195,21 @@ public class JSqrlServer {
                 //Disable the user's account
                 userService.disableSqrlUser(identityKey);
             } else if (command == SqrlCommand.REMOVE && sqrlEnabled) {
-                //Remove the user's account
-                verifyUnlockRequestSignature(request, sqrlUser.getVerifyUnlockKey(), verifier);
-                userService.removeSqrlUser(identityKey);
+                if (sqrlUser != null) {
+                    //Remove the user's account
+                    verifyUnlockRequestSignature(request, sqrlUser.getVerifyUnlockKey(), verifier);
+                    userService.removeSqrlUser(identityKey);
+                } else {
+                    tifs.add(TransactionInformationFlag.CLIENT_FAILURE);
+                }
             } else if (command == SqrlCommand.ENABLE) {
-                //Re-enable the user's account
-                verifyUnlockRequestSignature(request, sqrlUser.getVerifyUnlockKey(), verifier);
-                userService.enableSqrlUser(identityKey);
+                if (sqrlUser != null) {
+                    //Re-enable the user's account
+                    verifyUnlockRequestSignature(request, sqrlUser.getVerifyUnlockKey(), verifier);
+                    userService.enableSqrlUser(identityKey);
+                } else {
+                    tifs.add(TransactionInformationFlag.CLIENT_FAILURE);
+                }
             }
 
         }
@@ -223,7 +225,9 @@ public class JSqrlServer {
 
     }
 
-    private SqrlAuthResponse createResponse(String nut, String suk, TransactionInformationFlag... tifs) {
+    private SqrlAuthResponse createResponse(final String nut,
+                                            final String suk,
+                                            final TransactionInformationFlag... tifs) {
 
         return SqrlAuthResponse.builder()
                 .nut(nut)
@@ -234,7 +238,7 @@ public class JSqrlServer {
 
     }
 
-    private void verifyIdSignature(SqrlClientRequest request, Signature verifier) {
+    private void verifyIdSignature(final SqrlClientRequest request, final Signature verifier) {
         verifySqrlRequestSignature(
                 request,
                 verifier,
@@ -243,7 +247,9 @@ public class JSqrlServer {
                 "Unable to verify ID Signature");
     }
 
-    private void verifyUnlockRequestSignature(SqrlClientRequest request, String verifyUnlockKey, Signature verifier) {
+    private void verifyUnlockRequestSignature(final SqrlClientRequest request,
+                                              final String verifyUnlockKey,
+                                              final Signature verifier) {
         verifySqrlRequestSignature(
                 request,
                 verifier,
@@ -252,7 +258,8 @@ public class JSqrlServer {
                 "Unable to verify Unlock Request Signature");
     }
 
-    private void verifyPreviousIdSignature(SqrlClientRequest request, Signature verifier) {
+    private void verifyPreviousIdSignature(final SqrlClientRequest request,
+                                           final Signature verifier) {
         if (request.getPreviousIdentityKey() != null) {
             verifySqrlRequestSignature(
                     request,
@@ -263,7 +270,11 @@ public class JSqrlServer {
         }
     }
 
-    private void verifySqrlRequestSignature(SqrlClientRequest request, Signature verifier, byte[] key, byte[] signature, String errorMessage) {
+    private void verifySqrlRequestSignature(final SqrlClientRequest request,
+                                            final Signature verifier,
+                                            final byte[] key,
+                                            final byte[] signature,
+                                            final String errorMessage) {
         byte[] requestMessage = (request.getClient() + request.getServer()).getBytes();
         try {
             if (!verifyEdDSASignature(verifier, key, requestMessage, signature)) {
@@ -274,10 +285,10 @@ public class JSqrlServer {
         }
     }
 
-    private Boolean verifyEdDSASignature(Signature verifier,
-                                         byte[] key,
-                                         byte[] message,
-                                         byte[] signature) throws InvalidKeyException, SignatureException {
+    private Boolean verifyEdDSASignature(final Signature verifier,
+                                         final byte[] key,
+                                         final byte[] message,
+                                         final byte[] signature) throws InvalidKeyException, SignatureException {
         verifier.initVerify(new EdDSAPublicKey(new EdDSAPublicKeySpec(key, edDsaSpec)));
         verifier.update(message);
         return verifier.verify(signature);
